requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
gunicorn
jinja2
pytest
requests
httpx
pre-commit
black
python-multipart
pillow